requirements.txt
tensorflow==2.19.0
numpy==1.24.3
scikit-learn==1.2.2
joblib==1.3.2
pillow==9.5.0
opencv-python-headless==4.8.1.78
fastapi==0.95.2
uvicorn==0.22.0
python-multipart==0.0.6
wikipedia==1.4.0
h5py==3.9.0